Fix Radio Channel datamodel parameter
Issue: OFWR-1329
During an ApplySetting after setting a new channel, the wifi-agent calls wifi_reset or wifi_initRadio to reinitialize the radio.
It immediately calls CosaWifiReInitialize afterwards which ultimately calls wifi_getRadioChannel.
Since the radio reinitialization is done in the background, the newly configured channel is overwritten in the Radio struct by the call to wifi_getRadioChannel (which returns the running channel).

If we try to set the old channel afterwards, Radio_SetParamUlongValue will not configure it because the new channel equals the channel in the Radio struct.

Fix: Change CosaWifiReInitialize to ultimately call wifi_getRadioConfiguredChannel which returns the configured channel and not the running one.