prevent spaces being added to Wi-Fi passphrase using URL encoding ( OFW-351 )
[Description]
The issue was that the spaces are currently allowed in the WiFi passphrase.
A fix to validate that only ASCII characters between 33 and 126 are allowed in the passphrase has been added.

[Test Description]:
Try setting the Device.WiFi.AccessPoint.1.Security.X_COMCAST-COM_KeyPassphrase with string having space and any character which are outside 33 to 126
Should show error "CCSP_ERR_INVALID_PARAMETER_VALUE"

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)