fix channel is not set to manual after SPV on Device.WiFi.Radio.10100.AutoChannelEnable 5GHz ( OFW-1047 )
[Description]
Channel is not set to manual after SPV on Device.WiFi.Radio.{i}.AutoChannelEnable, this was for both 2.4 and 5 G radio
RCA: '0' channel was set to driver in case when the AutoChannelEnable was disabled instead of current channel.
Modified to to set the current running channel in case if the channel is not set by the user.

[Test Description]:
Set Device.WiFi.Radio.{i}.AutoChannelEnable to false and apply settings
Check AutoChannelEnable is having  the proper value is set.
Set Device.WiFi.Radio.{i}.Channel to valid value and Apply settings
Check AutoChannelEnable is disabled and channel is set to proper value

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)