fix init of size argument passed to COSAGetParamValueByPathName()
The COSAGetParamValueByPathName() function now checks the size value
passed in via its third argument, therefore the size needs to be
correct.

The COSAGetParamValueByPathName() function modifies the size value
passed in via its third argument, therefore if the varible holding
the size is reused by multiple calls to COSAGetParamValueByPathName()
it needs to be initialised before each one.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>