verify values set to X_LGI-COM_ReservedSSIDNames are single SSIDs
Commas are used as separators for the lists of reserved SSIDs,
therefore we need to enforce a rule that no reserved SSID can contain
a comma.

This also protects against attempts to set a list of SSIDs in one SPV
(which is not supported - reserved SSIDs need to be added one at a
time).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>