fix Device.WiFi.AccessPoint.1.AssociatedDevice.1.Active is showing True After wifi Disconnect ( OFW-1597 )
[Description]
Device.WiFi.AccessPoint.1.AssociatedDevice.{i}. should disappear when the client is disconnected.
This was not happening in our case as associated device count was not filled in by the apis.c when no clients are connected.
Added changes to fill the associated device count even when no clients are connected

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)