move platform specific 2 <-> 3 char country code conversion out of generic code ( OFW-1741 )
Part of fix for invalid parameters or instances found trying to verify CWMP Profile WiFiRadio:1 on Mv2+

[Description]
The datamodel was returning 4 char string, definition says its should be only 3 in broadband forum.
No need of additional "I" in Mv2+ as driver already return a proper value
Pushing this logic of appending 'I' to HAL of Mv1.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)