Only send LMLite notification for newly connected STA and not all STAs ( OFW-1945 )
This fixes a crash that occurs when testing a large number of
successive station associations using Veriwave. The crash occured
because of the large number of successive datamodel sets. This is
because a single station connection caused a notification to be sent
for each connected station.

Fix: Only send the notification for the newly connected station.