align X_LGI-COM_EnhancedACS to requirements ( OFW-1897 )
Make a get of Device.WiFi.Radio.1.X_LGI-COM_EnhancedACS.Weight.Channel
return all 13 channels, with only channels 1, 6 and 11 being non-zero.
For Radio 2, always return the configured value for channels 132, 136
and 140 without checking the bandwidth. Also remove the check for DFS
channels.

Change the size of the ChannelWeights array to 28 because the Broadcom
wifi_getRadioChannelWeights and wifi_setRadioChannelWeights expect an
array of at least that size.