allow a channel to be set after opensync moved away from it ( OFW-2010 )
After setting a manual channel from the UI then enabling opensync,
opensync can modify the channel. If we set the old channel in the UI
again while disabling opensync at the same time, the old channel won't
be set because the wifi-agent doesn't see a configuration change.
Fix: Allow sets to Channel to trigger a radio reset when the running
channel doesn't match the stored channel