fix 2.4 GHz Wireless MAC filtering accepts more than 16 devices ( OFW-1857 )
BUG-DESCRIPTION
As per the requirement the total of 32 devices are allowed to be added in macfiltering table

FIX DESCRIPTION:
While adding entry to the list we checked the number of devices already there in Access Point 1 , 2 for private network. With this we checked the total devices then only allowed device entry.

TESTING PERFORMED:
Executed the below commands on console to check.

  dmcli eRT addtable Device.WiFi.AccessPoint.1.X_CISCO_COM_MacFilterTable.
  dmcli eRT setv Device.WiFi.AccessPoint.1.X_CISCO_COM_MacFilterTable.1.DeviceName string "enter_name"
  dmcli eRT setv Device.WiFi.AccessPoint.1.X_CISCO_COM_MacFilterTable.1.MACAddress string "mac address".

These commands were executed total 32 times. Allowed access point here are 1,2
We may be able to add blank entries but while adding macaddress and device name if limit exceed it will delete the entry.