fix WiFi Network Name (SSID) broadcast is Disabled After reboot ( OFW-2337 CH7465SI-164 )
[Issue]
Because of dhcpv4 boot optimization the earlier logic of disable and enable SSID advertisement messages are not working.
It leads to SSIDs are not re enabled.

[Description]
Dhcpv4 is now started by 4 minutes of boot time, hence wifi can be started sooned with SSIDs enabled.

[Solution]
Remove the work around of disable SSID advertisement.

[Testing Performed]
Verify Wifi works after boot in all 3 modes with plume enabled.

Signed-off-by: Sidda(sihiregoudar@libertyglobal.com)