remove /nvram/TZ during factory reset on Puma6 Atom ( OFW-2477 CH7465SI-246 )
[Issue]
Device.Time.LocalTimeZone object not working as expected.

[Solution]
ToD server sends the correct time and an offest.One hour time difference occured because of adding the offset to the correct time.
Removed the offset to get the correct time.
Added code to export TZ value in atom to get the new timezone.
Modified the code to set the utc time in atom.

[Testing Performed]
1. Do a Factory Reset
2. Check the time in arm and atom via date and date -u command
3. Verified the time. Both arm and arm have the same time
4. Change the timezone to CET
   dmcli eRT setv Device.Time.LocalTimeZone string "CET-1CEST-2,M3.5.0/02:00:00,M10.5.0/03:00:00"
5. Verified the time.