temp fix: add support for parameter Device.WiFi.AccessPoint.{i}.AssociatedDevice.{i}.OperatingStandard ( OFW-3444 )
Needs to be reverted at rnl14, as the parameter can be aliased to
Device.WiFi.AccessPoint.{i}.AssociatedDevice.i.X_COMCAST-COM_OperatingStandard.

[Description]
This change is specifically for 1.3 only. The changes are to be dropped in 1.4 based releases.

Telenet is using the Device.WiFi.AccessPoint.{i}.AssociatedDevice.i.X_COMCAST-COM_OperatingStandard
parameter in their cloudUI and apparently need time to switch to the aliased parameter
Device.WiFi.AccessPoint.10101.AssociatedDevice.{i}.OperatingStandard.

The aliasing was originally done to be aligned with the DM published by the CPE requirements.

[Solution]
Alias manager functionality can be tweaked to address this specific concern.
However, that would then require extensive testing.

Since the ask is for the specific parameter and the parameter is a read only parameter,
the safest approach is to add the X_COMCAST-COM_OperatingStandard parameter in the
ccsp-wifi-agent and expose it over TR069.

[Testing Performed]
1. Boot up the gateway
2. Once the device is online, connect a WiFi client
3. Query the Device.WiFi.AccessPoint.{i}.AssociatedDevice.{i}.
3. Ensure that both X_COMCAST-COM_OperatingStandard and OperatingStandard DMs are
   present and filled in

Signed-off-by: Ajay Vishnu Bhat <abhat@libertyglobal.com>