fix CcspLmLite failes to receive notifications ( OFW-3477 )
No notification sent to LMLite if the wifi client is disconnected.

[Issue]
UI pages not loading on disconnection of WiFi clients.

[Description]
Currently if the associated devices are NULL the later step of sending notification to LMLite is not executed.

[Solution]
Do not proceed to send CcspLmLite notification if the associated device is NULL and count is greater than 0.
This avoids NULL pointer dereference and allows the lmlite notfication during disconnections.

[Testing]
- Connect the WiFi client verify Device.Hosts.Host.
- Disconnect Wifi Client and  Device.Hosts.Host., now Active paramater should show false.
- Verify the all the UI pages are loading (including DHCP).

Signed-off-by: Sidda(sihiregoudar@libertyglobal.com)