fix InterimInterval value should be configurable between 60 and above ( OFW-3701 CH7465SI-526 )
The value zero used to disable the fucntionality.

SPV on Device.WiFi.AccessPoint.[i].Accounting.InterimInterval doesn't return fault code as 9007 with an invalid range value.

BUG DESCRIPTION:
SPV of Device.WiFi.AccessPoint.[i].Accounting.InterimInterval is allowing the value less than 60, As per the MVXReq it should not lessthan 60.

FIX DESCRIPTION:
Changes are made as per MVXReq to return the falut code as 9007 with an invalid parameter value if the intermInterval is lessthan 60.

TESTING PERFORMED:
  1. Bring up the DUT.
  2. Enable the ACS
  3. Perform SPV on "Device.WiFi.AccessPoint.[i].Accounting.InterimInterval" with the value lessthan 60" and observed falut code as 9007 with an invalid parameter value.
  4. And verifed the SPV with different value greaterthan 60, and observed it is successfull.
  5. The above behaviour is also validated with dmcli.

Signed-off-by: Narayana DVVS <ndvvs.contractor@libertyglobal.com>