fix Radio reset is called during RF disconnection in Mv1 Device ( OFW-3853 )
Upstreamed the patch
sagemcom_0006_fix-bhaul-ssid-revert-after-configapply.patch.

[Issue]
If the gateway boots up with no pods claimed on the NOC, then the
backhaul SSIDs will be disabled and the wifi-agent will store that in
its internal config data structure during bootup.
If a pod is claimed afterwards from the NOC, then opensync will enable
the backhaul SSIDs through some HAL APIs.
When the RF cable is reconnected, the docsis boot file is downloaded and
the WiFi datamodel parameters included in the boot file are set.
At the end a WiFi ApplySetting is performed, which triggers a disable of
the backhaul SSIDs because of a mismatch between the SSID status and the
stored configuration.

[Description]
NA

[Solution]
Ignore changes to the backhaul SSIDs.
A similar change has been made for the PUMA7 platform.

[Testing performed]
Enable plume without claiming any pod to the gateway
Reboot the gateway
After the gateway has come up online on the NOC, claim a pod
Unplug the RF cable for a few minutes then reconnect it
Result:
The backhaul SSIDs should stay enabled and no radio reset triggered by opensync should occur