hal_version_3: fix Mismatch between HAL and dmcli values of RtsThresholdSupported [WIFI_HAL_VERSION_3] ( OFW-4213 )
BUG DESCRIPTION:
With  WIFI_HAL_VERSION_3 enabled, the Datemodel value of X_RDKCENTRAL-COM_RtsThresholdSupported
is not matching the value in the driver.
There is a missing HAL_API call to update the parameter under HAL_V3 flag.

FIX DESCRIPTION:
Added the call to HAL API while reading the radio configuration.

TESTING PERFORMED:
Tested and ensured that data model value for RtsThresholdSupported is matching with value in driver.

Upstream Candidate: Yes.

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)