temp_fix: modem broadcasts 2 different guest networks after factory reset causing multiple PATCH/POST 409 errors in C-Web ( OFW-4133 )
BUG DESCRIPTION:
Either of the password is not empty then leads to have different guest wifi passwords for 2.4 and 5GHz.
This leads to issue in C-APP C-CARE apps.

FIX DESCRIPTION:
set the same password for both the 2.4 and 5 GHz guest passwords, if either is empty.

TESTING PERFORMED:
1. Enable guest WiFi from UI
2. Change the password.
3. Do factory reset

rmcs: Change to be reverted once OEM implements proper fix.

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)