hal_version_3: fix SPV on "Device.WiFi.Radio.{i}.Channel" with invalid values is successful via ACS ( OFW-4398 F3896SI-1849 )
BUG DESCRIPTION:
When setting the new value to the channel,
validation is not happening whether the channel is valid or invalid when HAL v3 flag is enabled.

FIX DESCRIPTION:
Added the validation check for the channel
under SPV call to return error for invalid value.

TESTING PERFORMED:
1. Load the image with changes.
2. Login to the ACS
3. Perform SPV on "Device.WiFi.Radio.10000.Channel" and "Device.WiFi.Radio.10100.Channel" with invalid values.
       "Device.WiFi.Radio.10000.Channel" : 14
       "Device.WiFi.Radio.10100.Channel" : 37
4. Observed the SPV Fail with Faultcode 9007 - Invalid Parameter Value.
5. verified with dmcli also.

rmcs: Faultcode is expected to return as per OFW requirement, No
functional impact.

Upstream Candidate: No.