fix the private wifi not broadcasting after upgrade from bridge mode (OFW-4766  F3896SI-1979)
[Issue]
Private Wifi is not broadcasting after upgrading from 6.5.10 (bridge) to the FW 6.6.7(dual)

[Description]
Value from the hal api wifi_getApSsidAdvertisementEnable which is stored in the array AdvEnable during wifi
initilaize  is used to update the ssid advertisemnt enable parameter. While upgrading from bridge mode to dual mode,
the wifi initailize is called multiple times and the array which is filled initially is reset to 0.
This will disable SsidAdvertisement parameter and affect the boardcasting of private wifi.

[Solution]
The value from the hal api is used to update the ssid advertisemnet parameter.

[Testing]
1. Load the image to the FW: 6.5.10 and change the mode to bride via config file
2. Upgrade to the FW : 6.6.7 and change the provisioning mode from bridge to dual mode via config file while image gets downloaded.
3.Once DUT is online, check if the private wifi is broadcasting and wlan clients able to connect --> Connect one wlan client.
4. Private wifi should be broadcasting and wlan client should able to connect.

[Upstream Candidate] - No

 Signed-off-by: Sreeja S Kumar (srekumar.contractor@libertyglobal.com)