fix community WiFi SSID does not match what is set in bootfile (OFW-4769 F3896SI-2091)
BUG DESCRIPTION:
In the boot config file, the community WiFi SSID names are being set with the reserved SSID names which
then failed to set successfully.Even from ACS/dmcli , set operation with reserved SSID name fails.
In the ccsp-wifi-agent, there is a check while trying to set SSID if it belongs to thereserved SSID list.
If so, set operation returns FAILURE.But as per MVXREQ-1061 and MVXREQ-916, it is possible for
community WiFi having reserved SSID name.

FIX DESCRIPTION:
An additional check has been added to allow if the SSID instaces are Community WiFi,
the check for the presence of SSID in reserved SSID list is skipped.

TESTING PERFORMED:
Tested below scenarios:
1. SPV of community WiFi SSID with reserved SSID names ( e.g: "Ziggo", "Telenet").
   SPV is success.
2. SPV of Private/Guest WiFi SSID with reserved SSID names
   SPV fails.

[Upstream Candidate] - No

Signed-off-by: Keerthana V J(kvijayakumarjalajaku.contractor@libertyglobal.com)