fix private network broadcast status turns on after sometime when the broadcast is selected as "No" in UI ( OFW-4582 F3896SI-1999 )
[Issue]
Broadcast status of private network is automatically turned on after sometime when the broadcast is selected as "No" in UI

[Description]
Wifi agent is not sending notification as the comparison of the new vap paramter and the hal parameter fails.
When the Opensync recalculating the VIF config, it init the VIF state and then the "ssid_broadcast" value in VIF state table has
been changed becauseof the HAL api wifi_createVAP changed the SSIDAdvertisementEnabled nvram value.
This new nvram value is updated in the WifiAp->AP.Cfg. SSIDAdvertisementEnabled parameter from the get ssid param api.
During radio apply the comparison is done with vap value and the updated AP.Cfg. SSIDAdvertisementEnabled.
As both has the updated value notification is not send.

[Solution]
Updated the logic to by removing the updation of the cfg SSIDAdvertisementEnabled  parameter from get api.
This structure variable is updating once the raio apply is done.

[Testing]
1. Factory reset the DUT using SNMP
2. Login to GUI-> Change SSID name for both 2.4G and 5G private network.
3. Select broadcast status as "NO" for both bands in UI at the same time .
4. Verify the settings are saved in UI and SSID should not broadcast in WiFi clients
5. Do refresh or wait for 5 mins,to check the settings are retained in UI

Upstreamable : Yes

Signed-off-by: Sreeja S Kumar (srekumar.contractor@libertyglobal.com)