fix guest network not working as expected when either 2.4GHz or 5GHz radio is disabled ( F3896SI-1884 OFW-4436 )
[Issue]
Guest network not working as expected when either 2.4GHz or 5GHz radio is Disabled

[Description]
As part of halv3 implementation, the guest interface(brlan7) is disabled if any one of the radio is disabled.
Because of this brlan7 loses ip.

[Solution]
Updated the logic to disable the guest interface if both the radios are disabled.
The guest interface will be enabled and also have ip if any one of the channel is enabled.

[Testing]
1.Login to GUI.
2.Enable the Guest Network (Advanced settings -> Wireless -> Guest Network).
3.Make sure that brlan7 will be up and has IP.
4.Disable either 2.4GHz or 5GHz radio from GUI (Advanced settings -> Wireless -> Wireless signal).
5.Connect clients to the Guest WiFi SSID.

Upstreamable : Yes

Signed-off-by: Sreeja S Kumar (srekumar.contractor@libertyglobal.com)