fix brlan7 interface is getting ipv4 address in modem mode ( OFW-4714 F3896SI-2064 )
BUG DESCRIPTION:
brlan7 is getting ipv4 address in modem mode

ROOT CAUSE:
The brlan7 interface is coming up if the guest wifi is enabled in bridgemode.

FIX DESCRIPTION:
Fix is done to bring brlan7 interface up only if the mode is not bridge.
As part of upstreaming the patch [sagemcom_0012_fix_guest_wifi_enabling.patch]
restructuring of code is done by adding the macro.
As the macro usuage is incorrect the brlan7 is enabled again.
Updated the code to take the changes from the patch as such.

TESTING PERFORMED:
1.Flash latest release image
2.Enable plume in bootfile, provision in dualstack.
3.Enable Guest network from UI
4.Access the UI and change to modem mode from UI
5.Once reboot completes, check the brlan7 interface

Signed-off-by: Sreeja S Kumar (srekumar.contractor@libertyglobal.com)