fix radio getting disabled after moving from bridge to router mode part 2 ( OFW-5797 CH7465SI-1078 )
[BUG DESCRITPION]
In CBN MV1, if we move from legacy CBN to OFW in bridge mode, then after moving to router mode
the radio status is showing is disabled .

[FIX DESCIPTION]
In some corner case scenarios, especially if we place the InitializationMode value at the bottom of the boot file,
the process CosaDmlWiFiSetForceDisableWiFiRadio get executed before creation of migration_to_mng variable.
Because of this radioStatus is not set properly, so added the condition for checking the migration_to_mng file also.
By the time unlinking of file happens from ssp_main.c variable migration_to_mng become true.

[TESTING PERFORMED]
1.Upgrade from CBN legacy bridge mode to OFW bridge mode.
2.Change the mode from bridge to router mode.
3.Verified that in router mode, radio status is showing enabled.