process_monitor_atom.sh: ensure busybox pidof is used ( and not the procps version )
The busybox and procps versions of pidof behave differently when used
to determine the PID of a script (the procps version requires the -x
command line option, which the busybox version neither requires or
supports).

To ensure consistent behaviour regardless of whether the procps
version of pidof is installed, always run the busybox version
explicitly (ie run "busybox pidof" instead of "pidof").

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>