fix UI-Gateway takes longer to set apply changes in wireless signal page ( OFW-6049 )
[Description]
The UI sometimes takes a long time to apply changes in the Wireless
signal page. A radio reset is performed sometimes even though only the
other radio has been modified.
These extra resets would depend on the random order in which the UI
would set the parameters and call ApplySettings for each radio.
An extra reset also happened if the 2.4GHz radio stayed disabled.

[Solution]
On ApplySetting, only allow the radio being modified to trigger a radio reset.

[Testing performed]
Navigate to the Wireless Signal page on the UI
Disable Plume (channel optimization)
Disable/enable the 2.4GHz and 5GHz radios multiples, sometimes one radio at a time and sometimes both at once
Verify from the WiFi logs and the output of ps that only the modified radio is reset